.gitignore: Revert decision to ignore .ps, .eps, .pdf
In 05060274f6df1b62b1f5ada31583e07595472da7, we imported the gitignore
that we are now currently using. However, we made a slight modification
to it: ignore .ps, .eps, .pdf.

We ignore them because they are the extensions for output files, which
should not be pushed to the repository.

We also acknowledged that some of the input files will also have these
extensions, and input files should be pushed to the repository because
they are legitimate files. We advised users to push these files manually
in such a situation.

However, users may forget to push these input files because it is not
reflected in 'git status'. It is easy to forget especially if we push a
lot of such input files to the repository.

Such input files might be unrecoverable if we do not push it and
subsequently lost them.

Let's be safer and un-ignore .ps, .eps, .pdf, so that the user will
realise that these files exist in the local copy but are not pushed to
the repository.

With this change, users will have to manually delete the output files.
That is a little inconvenience that we trade, for forgetting to push
input files, which can be problematic in the long run since these files
may be unrecoverable.